Moved container implementations to runtime library.
