+++ include/neso_particles/containers/sym_vector_pointer_cache.hpp
@@ -237,5 +237,8 @@ public:
   }
 };
 
+extern template class SymVectorPointerCache<REAL>;
+extern template class SymVectorPointerCache<INT>;
+
 } // namespace NESO::Particles
 #endif
