+++ src/containers/cell_dat.cpp
@@ -1,0 +1,8 @@
+#include <neso_particles/containers/cell_dat.hpp>
+
+namespace NESO::Particles {
+
+template class CellDat<REAL>;
+template class CellDat<INT>;
+
+} // namespace NESO::Particles
