+++ src/containers/cell_dat_const.cpp
@@ -1,0 +1,9 @@
+#include <neso_particles/containers/cell_dat_const.hpp>
+
+namespace NESO::Particles {
+
+template class CellDatConst<REAL>;
+template class CellDatConst<INT>;
+template class CellDatConst<int>;
+
+} // namespace NESO::Particles
