+++ src/containers/global_array.cpp
@@ -1,0 +1,9 @@
+#include <neso_particles/containers/global_array.hpp>
+
+namespace NESO::Particles {
+
+template class GlobalArray<REAL>;
+template class GlobalArray<INT>;
+template class GlobalArray<int>;
+
+} // namespace NESO::Particles
