+++ src/containers/local_array.cpp
@@ -1,0 +1,7 @@
+#include <neso_particles/containers/local_array.hpp>
+
+namespace NESO::Particles {
+template class LocalArray<REAL>;
+template class LocalArray<INT>;
+template class LocalArray<int>;
+} // namespace NESO::Particles
