+++ src/containers/sym_vector.cpp
@@ -1,0 +1,7 @@
+#include <neso_particles/containers/sym_vector.hpp>
+#include <neso_particles/containers/sym_vector_impl.hpp>
+
+namespace NESO::Particles {
+template class SymVector<REAL>;
+template class SymVector<INT>;
+} // namespace NESO::Particles
