+++ src/containers/sym_vector_pointer_cache.cpp
@@ -1,0 +1,6 @@
+#include <neso_particles/containers/sym_vector_pointer_cache.hpp>
+
+namespace NESO::Particles {
+template class SymVectorPointerCache<REAL>;
+template class SymVectorPointerCache<INT>;
+} // namespace NESO::Particles
